api-gateaway/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
httpx>=0.27.0
python-jose[cryptography]>=3.3.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
